requirements.txt
certifi==2016.2.28
cycler==0.10.0
matplotlib==2.1.0
numpy==1.12.1
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.1
scipy==1.0.0
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.1.0
transforms3d==0.3.1
Werkzeug==0.12.2
wincertstore==0.2
